webhook: extract the main code to pkg
