+++ pkg/webhook/webhook.go
@@ -1,0 +1,241 @@
+// Copyright © 2021 Banzai Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package webhook
+
+import (
+	"context"
+	"crypto/x509"
+	"io/ioutil"
+	"net/http"
+	"os"
+
+	"emperror.dev/errors"
+	vaultapi "github.com/hashicorp/vault/api"
+	"github.com/sirupsen/logrus"
+	"github.com/slok/kubewebhook/v2/pkg/model"
+	"github.com/slok/kubewebhook/v2/pkg/webhook/mutating"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/client-go/kubernetes"
+	logrusadapter "logur.dev/adapter/logrus"
+
+	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
+)
+
+type MutatingWebhook struct {
+	k8sClient kubernetes.Interface
+	namespace string
+	registry  ImageRegistry
+	logger    *logrus.Entry
+}
+
+func (mw *MutatingWebhook) VaultSecretsMutator(ctx context.Context, ar *model.AdmissionReview, obj metav1.Object) (*mutating.MutatorResult, error) {
+	vaultConfig := parseVaultConfig(obj)
+
+	if vaultConfig.Skip {
+		return &mutating.MutatorResult{}, nil
+	}
+
+	switch v := obj.(type) {
+	case *corev1.Pod:
+		return &mutating.MutatorResult{MutatedObject: v}, mw.mutatePod(ctx, v, vaultConfig, ar.Namespace, ar.DryRun)
+
+	case *corev1.Secret:
+		return &mutating.MutatorResult{MutatedObject: v}, mw.mutateSecret(v, vaultConfig)
+
+	case *corev1.ConfigMap:
+		return &mutating.MutatorResult{MutatedObject: v}, mw.mutateConfigMap(v, vaultConfig)
+
+	case *unstructured.Unstructured:
+		return &mutating.MutatorResult{MutatedObject: v}, mw.mutateObject(v, vaultConfig)
+
+	default:
+		return &mutating.MutatorResult{}, nil
+	}
+}
+
+func (mw *MutatingWebhook) getDataFromConfigmap(cmName string, ns string) (map[string]string, error) {
+	configMap, err := mw.k8sClient.CoreV1().ConfigMaps(ns).Get(context.Background(), cmName, metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+	return configMap.Data, nil
+}
+
+func (mw *MutatingWebhook) getDataFromSecret(secretName string, ns string) (map[string][]byte, error) {
+	secret, err := mw.k8sClient.CoreV1().Secrets(ns).Get(context.Background(), secretName, metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+	return secret.Data, nil
+}
+
+func (mw *MutatingWebhook) lookForEnvFrom(envFrom []corev1.EnvFromSource, ns string) ([]corev1.EnvVar, error) {
+	var envVars []corev1.EnvVar
+
+	for _, ef := range envFrom {
+		if ef.ConfigMapRef != nil {
+			data, err := mw.getDataFromConfigmap(ef.ConfigMapRef.Name, ns)
+			if err != nil {
+				if apierrors.IsNotFound(err) || (ef.ConfigMapRef.Optional != nil && *ef.ConfigMapRef.Optional) {
+					continue
+				} else {
+					return envVars, err
+				}
+			}
+			for key, value := range data {
+				if hasVaultPrefix(value) {
+					envFromCM := corev1.EnvVar{
+						Name:  key,
+						Value: value,
+					}
+					envVars = append(envVars, envFromCM)
+				}
+			}
+		}
+		if ef.SecretRef != nil {
+			data, err := mw.getDataFromSecret(ef.SecretRef.Name, ns)
+			if err != nil {
+				if apierrors.IsNotFound(err) || (ef.SecretRef.Optional != nil && *ef.SecretRef.Optional) {
+					continue
+				} else {
+					return envVars, err
+				}
+			}
+			for key, value := range data {
+				if hasVaultPrefix(string(value)) {
+					envFromSec := corev1.EnvVar{
+						Name:  key,
+						Value: string(value),
+					}
+					envVars = append(envVars, envFromSec)
+				}
+			}
+		}
+	}
+	return envVars, nil
+}
+
+func (mw *MutatingWebhook) lookForValueFrom(env corev1.EnvVar, ns string) (*corev1.EnvVar, error) {
+	if env.ValueFrom.ConfigMapKeyRef != nil {
+		data, err := mw.getDataFromConfigmap(env.ValueFrom.ConfigMapKeyRef.Name, ns)
+		if err != nil {
+			if apierrors.IsNotFound(err) {
+				return nil, nil
+			}
+			return nil, err
+		}
+		if hasVaultPrefix(data[env.ValueFrom.ConfigMapKeyRef.Key]) {
+			fromCM := corev1.EnvVar{
+				Name:  env.Name,
+				Value: data[env.ValueFrom.ConfigMapKeyRef.Key],
+			}
+			return &fromCM, nil
+		}
+	}
+	if env.ValueFrom.SecretKeyRef != nil {
+		data, err := mw.getDataFromSecret(env.ValueFrom.SecretKeyRef.Name, ns)
+		if err != nil {
+			if apierrors.IsNotFound(err) {
+				return nil, nil
+			}
+			return nil, err
+		}
+		if hasVaultPrefix(string(data[env.ValueFrom.SecretKeyRef.Key])) {
+			fromSecret := corev1.EnvVar{
+				Name:  env.Name,
+				Value: string(data[env.ValueFrom.SecretKeyRef.Key]),
+			}
+			return &fromSecret, nil
+		}
+	}
+	return nil, nil
+}
+
+func (mw *MutatingWebhook) newVaultClient(vaultConfig VaultConfig) (*vault.Client, error) {
+	clientConfig := vaultapi.DefaultConfig()
+	if clientConfig.Error != nil {
+		return nil, clientConfig.Error
+	}
+
+	clientConfig.Address = vaultConfig.Addr
+
+	tlsConfig := vaultapi.TLSConfig{Insecure: vaultConfig.SkipVerify}
+	err := clientConfig.ConfigureTLS(&tlsConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	if vaultConfig.TLSSecret != "" {
+		tlsSecret, err := mw.k8sClient.CoreV1().Secrets(mw.namespace).Get(
+			context.Background(),
+			vaultConfig.TLSSecret,
+			metav1.GetOptions{},
+		)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to read Vault TLS Secret")
+		}
+
+		clientTLSConfig := clientConfig.HttpClient.Transport.(*http.Transport).TLSClientConfig
+
+		pool := x509.NewCertPool()
+
+		ok := pool.AppendCertsFromPEM(tlsSecret.Data["ca.crt"])
+		if !ok {
+			return nil, errors.Errorf("error loading Vault CA PEM from TLS Secret: %s", tlsSecret.Name)
+		}
+
+		clientTLSConfig.RootCAs = pool
+	}
+
+	return vault.NewClientFromConfig(
+		clientConfig,
+		vault.ClientRole(vaultConfig.Role),
+		vault.ClientAuthPath(vaultConfig.Path),
+		vault.ClientAuthMethod(vaultConfig.AuthMethod),
+		vault.ClientLogger(logrusadapter.NewFromEntry(mw.logger)),
+	)
+}
+
+func (mw *MutatingWebhook) ServeMetrics(addr string, handler http.Handler) {
+	mw.logger.Infof("Telemetry on http://%s", addr)
+
+	mux := http.NewServeMux()
+	mux.Handle("/metrics", handler)
+	err := http.ListenAndServe(addr, mux)
+	if err != nil {
+		mw.logger.Fatalf("error serving telemetry: %s", err)
+	}
+}
+
+func NewMutatingWebhook(logger *logrus.Entry, k8sClient kubernetes.Interface) (*MutatingWebhook, error) {
+	namespace := os.Getenv("KUBERNETES_NAMESPACE") // only for kurun
+	if namespace == "" {
+		namespaceBytes, err := ioutil.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
+		if err != nil {
+			return nil, errors.Wrap(err, "error reading k8s namespace")
+		}
+		namespace = string(namespaceBytes)
+	}
+
+	return &MutatingWebhook{
+		k8sClient: k8sClient,
+		namespace: namespace,
+		registry:  NewRegistry(),
+		logger:    logger,
+	}, nil
+}
